[Step Function] 6.2 Refactor Lambda tag code (#156)
