+++ serverless/test/lambda/tags.spec.ts
@@ -1,6 +1,7 @@
 import { LambdaConfigLoader } from "../../src/lambda/env";
 import { RuntimeType, LambdaFunction } from "../../src/lambda/layer";
-import { addDDTags, addMacroTag, addCDKTag, addSAMTag } from "../../src/lambda/tags";
+import { addCDKTag, addSAMTag } from "../../src/lambda/tags";
+import { addDDTags, addMacroTag } from "../../src/common/tags";
 
 function mockLambdaFunction(tags: { Key: string; Value: string }[]) {
   return {
@@ -33,14 +34,14 @@ describe("addDDTags", () => {
       { Key: "team", Value: "serverless" },
     ];
     const lambda = mockLambdaFunction(existingTags);
-    addDDTags([lambda], config);
+    addDDTags(lambda, config);
 
     expect(lambda.properties.Tags).toEqual([...existingTags, { Key: "project", Value: "marvel" }]);
   });
 
   it("does not add tags if provided config doesn't have tags", () => {
     const lambda = mockLambdaFunction([]);
-    addDDTags([lambda], configLoader.defaultConfiguration);
+    addDDTags(lambda, configLoader.defaultConfiguration);
 
     expect(lambda.properties.Tags).toEqual([]);
   });
@@ -64,7 +65,7 @@ describe("addDDTags", () => {
         service: "my-service",
       };
       const lambda = mockLambdaFunction([]);
-      addDDTags([lambda], config);
+      addDDTags(lambda, config);
 
       expect(lambda.properties.Tags).toEqual([
         { Key: "service", Value: "my-service" },
@@ -82,7 +83,7 @@ describe("addDDTags", () => {
       tags: "team:avengers,project:marvel",
     };
     const lambda = mockLambdaFunction([]);
-    addDDTags([lambda], config);
+    addDDTags(lambda, config);
 
     expect(lambda.properties.Tags).toEqual([
       { Key: "service", Value: "my-service" },
@@ -105,7 +106,7 @@ describe("addDDTags", () => {
       { Key: "project", Value: "lambda" },
     ];
     const lambda = mockLambdaFunction(existingTags);
-    addDDTags([lambda], config);
+    addDDTags(lambda, config);
 
     expect(lambda.properties.Tags).toEqual([
       { Key: "env", Value: "dev" },
@@ -121,14 +122,14 @@ describe("addMacroTag", () => {
   it("does not update tags if no version is passed in", () => {
     const existingTags = [{ Key: "band", Value: "ironmaiden" }];
     const lambda = mockLambdaFunction(existingTags);
-    addMacroTag([lambda], undefined);
+    addMacroTag(lambda, undefined);
 
     expect(lambda.properties.Tags).toEqual(existingTags);
   });
 
   it("creates tags property if needed", () => {
     const lambda = mockLambdaFunction([]);
-    addMacroTag([lambda], "6.6.6");
+    addMacroTag(lambda, "6.6.6");
 
     expect(lambda.properties.Tags).toEqual([{ Key: "dd_sls_macro", Value: "v6.6.6" }]);
   });
@@ -136,7 +137,7 @@ describe("addMacroTag", () => {
   it("appends version tag if needed", () => {
     const existingTags = [{ Key: "env", Value: "dev" }];
     const lambda = mockLambdaFunction(existingTags);
-    addMacroTag([lambda], "6.6.6");
+    addMacroTag(lambda, "6.6.6");
 
     expect(lambda.properties.Tags).toEqual([
       { Key: "env", Value: "dev" },
